port all chain impls to the new api
