+++ chain/bitcoin/bitcoin_test.go
@@ -13,7 +13,6 @@ import (
 	"github.com/renproject/multichain/api/address"
 	"github.com/renproject/multichain/api/utxo"
 	"github.com/renproject/multichain/chain/bitcoin"
-	"github.com/renproject/multichain/compat/bitcoincompat"
 	"github.com/renproject/pack"
 
 	. "github.com/onsi/ginkgo"
@@ -112,7 +111,7 @@ var _ = Describe("Bitcoin", func() {
 				Expect(err).ToNot(HaveOccurred())
 				err = client.SubmitTx(context.Background(), tx)
 				Expect(err).ToNot(HaveOccurred())
-				log.Printf("TXID               %v")
+				log.Printf("TXID               %v", txHash)
 
 				for {
 					// Loop until the transaction has at least a few
@@ -135,132 +134,6 @@ var _ = Describe("Bitcoin", func() {
 				output2, _, err = client.Output(context.Background(), output.Outpoint)
 				Expect(err).ToNot(HaveOccurred())
 				Expect(reflect.DeepEqual(output, output2)).To(BeTrue())
-			})
-		})
-
-		Context("when sending BTC to a script", func() {
-			It("should work", func() {
-				// Private key
-				pkEnv := os.Getenv("BITCOIN_PK")
-				if pkEnv == "" {
-					panic("BITCOIN_PK is undefined")
-				}
-				wif, err := btcutil.DecodeWIF(pkEnv)
-				Expect(err).ToNot(HaveOccurred())
-
-				// Client
-				client := bitcoin.NewClient(bitcoin.DefaultClientOptions())
-
-				// Script
-				gpubkey := wif.PrivKey.PubKey().SerializeCompressed()
-				ghash := [32]byte{}
-				gscript, err := bitcoincompat.GatewayScript(gpubkey, ghash)
-				Expect(err).ToNot(HaveOccurred())
-				gaddr, err := btcutil.NewAddressScriptHash(gscript, &chaincfg.RegressionNetParams)
-				Expect(err).ToNot(HaveOccurred())
-
-				// PKH
-				pkh, err := btcutil.NewAddressPubKeyHash(btcutil.Hash160(wif.PrivKey.PubKey().SerializeCompressed()), &chaincfg.RegressionNetParams)
-				Expect(err).ToNot(HaveOccurred())
-				log.Printf("PKH %v", pkh.EncodeAddress())
-
-				scriptOutput := utxo.Output{}
-				{
-					// Outputs
-					outputs, err := client.UnspentOutputs(context.Background(), 0, 999999999, address.Address(pkh.EncodeAddress()))
-					Expect(err).ToNot(HaveOccurred())
-					Expect(len(outputs)).To(BeNumerically(">", 0))
-					output := outputs[0]
-
-					// Input, recipients, and transaction
-					inputs := []utxo.Input{
-						{Output: output},
-					}
-					recipients := []utxo.Recipient{
-						{To: address.Address(gaddr.EncodeAddress()), Value: pack.NewU256FromU64(pack.NewU64(output.Value.Int().Uint64() - 1000))},
-					}
-					tx, err := bitcoin.NewTxBuilder(&chaincfg.RegressionNetParams).BuildTx(inputs, recipients)
-					Expect(err).ToNot(HaveOccurred())
-
-					// Sign
-					sighashes, err := tx.Sighashes()
-					signatures := make([]pack.Bytes65, len(sighashes))
-					Expect(err).ToNot(HaveOccurred())
-					for i := range sighashes {
-						hash := id.Hash(sighashes[i])
-						privKey := (*id.PrivKey)(wif.PrivKey)
-						signature, err := privKey.Sign(&hash)
-						Expect(err).ToNot(HaveOccurred())
-						signatures[i] = pack.NewBytes65(signature)
-					}
-					Expect(tx.Sign(signatures, pack.NewBytes(wif.SerializePubKey()))).To(Succeed())
-
-					// Submit
-					txHash, err := tx.Hash()
-					Expect(err).ToNot(HaveOccurred())
-					err = client.SubmitTx(context.Background(), tx)
-					Expect(err).ToNot(HaveOccurred())
-
-					for {
-						// Confirm
-						confs, err := client.Confirmations(context.Background(), txHash)
-						Expect(err).ToNot(HaveOccurred())
-						log.Printf("%v has %v/3 confirmations", txHash, confs)
-						if confs >= 1 {
-							break
-						}
-						time.Sleep(time.Second)
-					}
-
-					outputs, err = tx.Outputs()
-					Expect(err).ToNot(HaveOccurred())
-					Expect(outputs).To(HaveLen(1))
-					scriptOutput = outputs[0]
-				}
-
-				{
-					// Input, recipients, and transaction
-					inputSigScript, err := bitcoincompat.GatewayScript(gpubkey, ghash)
-					Expect(err).ToNot(HaveOccurred())
-					inputs := []utxo.Input{
-						{Output: scriptOutput, SigScript: inputSigScript},
-					}
-					recipients := []utxo.Recipient{
-						{To: address.Address(pkh.EncodeAddress()), Value: pack.NewU256FromU64(pack.NewU64(scriptOutput.Value.Int().Uint64() - 1000))},
-					}
-					tx, err := bitcoin.NewTxBuilder(&chaincfg.RegressionNetParams).BuildTx(inputs, recipients)
-					Expect(err).ToNot(HaveOccurred())
-
-					// Sign
-					sighashes, err := tx.Sighashes()
-					signatures := make([]pack.Bytes65, len(sighashes))
-					Expect(err).ToNot(HaveOccurred())
-					for i := range sighashes {
-						hash := id.Hash(sighashes[i])
-						privKey := (*id.PrivKey)(wif.PrivKey)
-						signature, err := privKey.Sign(&hash)
-						Expect(err).ToNot(HaveOccurred())
-						signatures[i] = pack.NewBytes65(signature)
-					}
-					Expect(tx.Sign(signatures, pack.NewBytes(wif.SerializePubKey()))).To(Succeed())
-
-					// Submit
-					txHash, err := tx.Hash()
-					Expect(err).ToNot(HaveOccurred())
-					err = client.SubmitTx(context.Background(), tx)
-					Expect(err).ToNot(HaveOccurred())
-
-					for {
-						// Confirm
-						confs, err := client.Confirmations(context.Background(), txHash)
-						Expect(err).ToNot(HaveOccurred())
-						log.Printf("%v has %v/3 confirmations", txHash, confs)
-						if confs >= 3 {
-							break
-						}
-						time.Sleep(time.Second)
-					}
-				}
 			})
 		})
 	})
