+++ chain/bitcoincash/bitcoincash_test.go
@@ -10,8 +10,9 @@ import (
 	"github.com/btcsuite/btcd/chaincfg"
 	"github.com/btcsuite/btcutil"
 	"github.com/renproject/id"
+	"github.com/renproject/multichain/api/address"
+	"github.com/renproject/multichain/api/utxo"
 	"github.com/renproject/multichain/chain/bitcoincash"
-	"github.com/renproject/multichain/compat/bitcoincompat"
 	"github.com/renproject/pack"
 
 	. "github.com/onsi/ginkgo"
@@ -41,8 +42,8 @@ var _ = Describe("Bitcoin Cash", func() {
 				log.Printf("PKH (uncompressed) %v", pkhAddrUncompressed.EncodeAddress())
 
 				// Setup the client and load the unspent transaction outputs.
-				client := bitcoincompat.NewClient(bitcoincompat.DefaultClientOptions().WithHost("http://127.0.0.1:19443"))
-				outputs, err := client.UnspentOutputs(context.Background(), 0, 999999999, pkhAddr)
+				client := bitcoincash.NewClient(bitcoincash.DefaultClientOptions())
+				outputs, err := client.UnspentOutputs(context.Background(), 0, 999999999, address.Address(pkhAddr.EncodeAddress()))
 				Expect(err).ToNot(HaveOccurred())
 				Expect(len(outputs)).To(BeNumerically(">", 0))
 				output := outputs[0]
@@ -56,17 +57,17 @@ var _ = Describe("Bitcoin Cash", func() {
 
 				// Build the transaction by consuming the outputs and spending
 				// them to a set of recipients.
-				inputs := []bitcoincompat.Input{
+				inputs := []utxo.Input{
 					{Output: output},
 				}
-				recipients := []bitcoincompat.Recipient{
+				recipients := []utxo.Recipient{
 					{
-						Address: pack.String(pkhAddr.EncodeAddress()),
-						Value:   pack.NewU64((output.Value.Uint64() - 1000) / 2),
+						To:    address.Address(pkhAddr.EncodeAddress()),
+						Value: pack.NewU256FromU64(pack.NewU64((output.Value.Int().Uint64() - 1000) / 2)),
 					},
 					{
-						Address: pack.String(pkhAddrUncompressed.EncodeAddress()),
-						Value:   pack.NewU64((output.Value.Uint64() - 1000) / 2),
+						To:    address.Address(pkhAddrUncompressed.EncodeAddress()),
+						Value: pack.NewU256FromU64(pack.NewU64((output.Value.Int().Uint64() - 1000) / 2)),
 					},
 				}
 				tx, err := bitcoincash.NewTxBuilder(&chaincfg.RegressionNetParams).BuildTx(inputs, recipients)
@@ -90,7 +91,9 @@ var _ = Describe("Bitcoin Cash", func() {
 
 				// Submit the transaction to the Bitcoin Cash node. Again, this
 				// should be running a la `./multichaindeploy`.
-				txHash, err := client.SubmitTx(context.Background(), tx)
+				txHash, err := tx.Hash()
+				Expect(err).ToNot(HaveOccurred())
+				err = client.SubmitTx(context.Background(), tx)
 				Expect(err).ToNot(HaveOccurred())
 				log.Printf("TXID               %v", txHash)
 
@@ -116,128 +119,6 @@ var _ = Describe("Bitcoin Cash", func() {
 				Expect(err).ToNot(HaveOccurred())
 				Expect(reflect.DeepEqual(output, output2)).To(BeTrue())
 			})
-		})
-	})
-
-	Context("when sending BCH to a script", func() {
-		It("should work", func() {
-			// Private key
-			pkEnv := os.Getenv("BITCOINCASH_PK")
-			if pkEnv == "" {
-				panic("BITCOINCASH_PK is undefined")
-			}
-			wif, err := btcutil.DecodeWIF(pkEnv)
-			Expect(err).ToNot(HaveOccurred())
-
-			// Client
-			client := bitcoincompat.NewClient(bitcoincompat.DefaultClientOptions().WithHost("http://127.0.0.1:19443"))
-
-			// Script
-			gpubkey := wif.PrivKey.PubKey().SerializeCompressed()
-			ghash := [32]byte{}
-			gscript, err := bitcoincompat.GatewayScript(gpubkey, ghash)
-			Expect(err).ToNot(HaveOccurred())
-			gaddr, err := bitcoincash.NewAddressScriptHash(gscript, &chaincfg.RegressionNetParams)
-			Expect(err).ToNot(HaveOccurred())
-
-			// PKH
-			pkh, err := bitcoincash.NewAddressPubKeyHash(btcutil.Hash160(wif.PrivKey.PubKey().SerializeCompressed()), &chaincfg.RegressionNetParams)
-			Expect(err).ToNot(HaveOccurred())
-			log.Printf("PKH %v", pkh.EncodeAddress())
-
-			scriptOutput := bitcoincompat.Output{}
-			{
-				// Outputs
-				outputs, err := client.UnspentOutputs(context.Background(), 0, 999999999, pkh)
-				Expect(err).ToNot(HaveOccurred())
-				Expect(len(outputs)).To(BeNumerically(">", 0))
-				output := outputs[0]
-
-				// Input, recipients, and transaction
-				inputs := []bitcoincompat.Input{
-					{Output: output},
-				}
-				recipients := []bitcoincompat.Recipient{
-					{Address: pack.String(gaddr.EncodeAddress()), Value: pack.NewU64(output.Value.Uint64() - 1000)},
-				}
-				tx, err := bitcoincash.NewTxBuilder(&chaincfg.RegressionNetParams).BuildTx(inputs, recipients)
-				Expect(err).ToNot(HaveOccurred())
-
-				// Sign
-				sighashes, err := tx.Sighashes()
-				signatures := make([]pack.Bytes65, len(sighashes))
-				Expect(err).ToNot(HaveOccurred())
-				for i := range sighashes {
-					hash := id.Hash(sighashes[i])
-					privKey := (*id.PrivKey)(wif.PrivKey)
-					signature, err := privKey.Sign(&hash)
-					Expect(err).ToNot(HaveOccurred())
-					signatures[i] = pack.NewBytes65(signature)
-				}
-				Expect(tx.Sign(signatures, pack.NewBytes(wif.SerializePubKey()))).To(Succeed())
-
-				// Submit
-				txHash, err := client.SubmitTx(context.Background(), tx)
-				Expect(err).ToNot(HaveOccurred())
-
-				for {
-					// Confirm
-					confs, err := client.Confirmations(context.Background(), txHash)
-					Expect(err).ToNot(HaveOccurred())
-					log.Printf("%v has %v/3 confirmations", txHash, confs)
-					if confs >= 1 {
-						break
-					}
-					time.Sleep(time.Second)
-				}
-
-				outputs, err = tx.Outputs()
-				Expect(err).ToNot(HaveOccurred())
-				Expect(outputs).To(HaveLen(1))
-				scriptOutput = outputs[0]
-			}
-
-			{
-				// Input, recipients, and transaction
-				inputSigScript, err := bitcoincompat.GatewayScript(gpubkey, ghash)
-				Expect(err).ToNot(HaveOccurred())
-				inputs := []bitcoincompat.Input{
-					{Output: scriptOutput, SigScript: inputSigScript},
-				}
-				recipients := []bitcoincompat.Recipient{
-					{Address: pack.String(pkh.EncodeAddress()), Value: pack.NewU64(scriptOutput.Value.Uint64() - 1000)},
-				}
-				tx, err := bitcoincash.NewTxBuilder(&chaincfg.RegressionNetParams).BuildTx(inputs, recipients)
-				Expect(err).ToNot(HaveOccurred())
-
-				// Sign
-				sighashes, err := tx.Sighashes()
-				signatures := make([]pack.Bytes65, len(sighashes))
-				Expect(err).ToNot(HaveOccurred())
-				for i := range sighashes {
-					hash := id.Hash(sighashes[i])
-					privKey := (*id.PrivKey)(wif.PrivKey)
-					signature, err := privKey.Sign(&hash)
-					Expect(err).ToNot(HaveOccurred())
-					signatures[i] = pack.NewBytes65(signature)
-				}
-				Expect(tx.Sign(signatures, pack.NewBytes(wif.SerializePubKey()))).To(Succeed())
-
-				// Submit
-				txHash, err := client.SubmitTx(context.Background(), tx)
-				Expect(err).ToNot(HaveOccurred())
-
-				for {
-					// Confirm
-					confs, err := client.Confirmations(context.Background(), txHash)
-					Expect(err).ToNot(HaveOccurred())
-					log.Printf("%v has %v/3 confirmations", txHash, confs)
-					if confs >= 3 {
-						break
-					}
-					time.Sleep(time.Second)
-				}
-			}
 		})
 	})
 })
