+++ chain/bitcoincash/utxo_test.go
@@ -1,0 +1,1 @@
+package bitcoincash_test
