+++ chain/dogecoin/address.go
@@ -2,6 +2,8 @@ package dogecoin
 
 import "github.com/renproject/multichain/chain/bitcoin"
 
-type AddressEncoder = bitcoin.AddressEncoder
-type AddressDecoder = bitcoin.AddressDecoder
-type AddressEncodeDecoder = bitcoin.AddressEncodeDecoder
+type (
+	AddressEncoder       = bitcoin.AddressEncoder
+	AddressDecoder       = bitcoin.AddressDecoder
+	AddressEncodeDecoder = bitcoin.AddressEncodeDecoder
+)
