+++ chain/dogecoin/dogecoin_test.go
@@ -9,6 +9,8 @@ import (
 
 	"github.com/btcsuite/btcutil"
 	"github.com/renproject/id"
+	"github.com/renproject/multichain/api/address"
+	"github.com/renproject/multichain/api/utxo"
 	"github.com/renproject/multichain/chain/dogecoin"
 	"github.com/renproject/pack"
 
@@ -39,8 +41,8 @@ var _ = Describe("Dogecoin", func() {
 				log.Printf("PKH (uncompressed) %v", pkhAddrUncompressed.EncodeAddress())
 
 				// Setup the client and load the unspent transaction outputs.
-				client := bitcoincompat.NewClient(bitcoincompat.DefaultClientOptions().WithHost("http://127.0.0.1:18332"))
-				outputs, err := client.UnspentOutputs(context.Background(), 0, 999999999, pkhAddr)
+				client := dogecoin.NewClient(dogecoin.DefaultClientOptions().WithHost("http://127.0.0.1:18332"))
+				outputs, err := client.UnspentOutputs(context.Background(), 0, 999999999, address.Address(pkhAddr.EncodeAddress()))
 				Expect(err).ToNot(HaveOccurred())
 				Expect(len(outputs)).To(BeNumerically(">", 0))
 				output := outputs[0]
@@ -54,17 +56,17 @@ var _ = Describe("Dogecoin", func() {
 
 				// Build the transaction by consuming the outputs and spending
 				// them to a set of recipients.
-				inputs := []bitcoincompat.Input{
+				inputs := []utxo.Input{
 					{Output: output},
 				}
-				recipients := []bitcoincompat.Recipient{
+				recipients := []utxo.Recipient{
 					{
-						Address: pack.String(pkhAddr.EncodeAddress()),
-						Value:   pack.NewU64((output.Value.Uint64() - 1000) / 2),
+						To:    address.Address(pkhAddr.EncodeAddress()),
+						Value: pack.NewU256FromU64(pack.NewU64((output.Value.Int().Uint64() - 1000) / 2)),
 					},
 					{
-						Address: pack.String(pkhAddrUncompressed.EncodeAddress()),
-						Value:   pack.NewU64((output.Value.Uint64() - 1000) / 2),
+						To:    address.Address(pkhAddrUncompressed.EncodeAddress()),
+						Value: pack.NewU256FromU64(pack.NewU64((output.Value.Int().Uint64() - 1000) / 2)),
 					},
 				}
 				tx, err := dogecoin.NewTxBuilder(&dogecoin.RegressionNetParams).BuildTx(inputs, recipients)
@@ -88,7 +90,9 @@ var _ = Describe("Dogecoin", func() {
 
 				// Submit the transaction to the Dogecoin node. Again, this
 				// should be running a la `./multichaindeploy`.
-				txHash, err := client.SubmitTx(context.Background(), tx)
+				txHash, err := tx.Hash()
+				Expect(err).ToNot(HaveOccurred())
+				err = client.SubmitTx(context.Background(), tx)
 				Expect(err).ToNot(HaveOccurred())
 				log.Printf("TXID               %v", txHash)
