+++ chain/terra/address.go
@@ -1,13 +1,15 @@
 package terra
 
-import (
-	"github.com/renproject/multichain/chain/cosmos"
-	"github.com/renproject/multichain/compat/cosmoscompat"
+import "github.com/renproject/multichain/chain/cosmos"
+
+type (
+	AddressDecoder       = cosmos.AddressDecoder
+	AddressEncoder       = cosmos.AddressEncoder
+	AddressEncodeDecoder = cosmos.AddressEncodeDecoder
 )
 
-// NewAddressDecoder returns an implementation of the address decoder interface
-// from the Cosmos Compat API, and exposes the functionality to decode strings
-// into addresses.
-func NewAddressDecoder() cosmoscompat.AddressDecoder {
-	return cosmos.NewAddressDecoder("terra")
-}
+var (
+	NewAddressDecoder      = cosmos.NewAddressDecoder
+	NewAddressEncoder      = cosmos.NewAddressEncoder
+	NewAddressEnodeDecoder = cosmos.NewAddressEncodeDecoder
+)
