+++ chain/zcash/address_test.go
@@ -1,0 +1,1 @@
+package zcash_test
