+++ chain/zcash/zcash_test.go
@@ -7,11 +7,11 @@ import (
 	"reflect"
 	"time"
 
-	"github.com/btcsuite/btcd/chaincfg"
 	"github.com/btcsuite/btcutil"
 	"github.com/renproject/id"
+	"github.com/renproject/multichain/api/address"
+	"github.com/renproject/multichain/api/utxo"
 	"github.com/renproject/multichain/chain/zcash"
-	"github.com/renproject/multichain/compat/bitcoincompat"
 	"github.com/renproject/pack"
 
 	. "github.com/onsi/ginkgo"
@@ -33,16 +33,16 @@ var _ = Describe("Zcash", func() {
 				Expect(err).ToNot(HaveOccurred())
 
 				// PKH
-				pkhAddr, err := zcash.NewAddressPubKeyHash(btcutil.Hash160(wif.PrivKey.PubKey().SerializeCompressed()), &chaincfg.RegressionNetParams)
+				pkhAddr, err := zcash.NewAddressPubKeyHash(btcutil.Hash160(wif.PrivKey.PubKey().SerializeCompressed()), &zcash.RegressionNetParams)
 				Expect(err).ToNot(HaveOccurred())
-				pkhAddrUncompressed, err := zcash.NewAddressPubKeyHash(btcutil.Hash160(wif.PrivKey.PubKey().SerializeUncompressed()), &chaincfg.RegressionNetParams)
+				pkhAddrUncompressed, err := zcash.NewAddressPubKeyHash(btcutil.Hash160(wif.PrivKey.PubKey().SerializeUncompressed()), &zcash.RegressionNetParams)
 				Expect(err).ToNot(HaveOccurred())
 				log.Printf("PKH                %v", pkhAddr.EncodeAddress())
 				log.Printf("PKH (uncompressed) %v", pkhAddrUncompressed.EncodeAddress())
 
 				// Setup the client and load the unspent transaction outputs.
-				client := bitcoincompat.NewClient(bitcoincompat.DefaultClientOptions().WithHost("http://127.0.0.1:18232"))
-				outputs, err := client.UnspentOutputs(context.Background(), 0, 999999999, pkhAddr)
+				client := zcash.NewClient(zcash.DefaultClientOptions())
+				outputs, err := client.UnspentOutputs(context.Background(), 0, 999999999, address.Address(pkhAddr.EncodeAddress()))
 				Expect(err).ToNot(HaveOccurred())
 				Expect(len(outputs)).To(BeNumerically(">", 0))
 				output := outputs[0]
@@ -56,20 +56,20 @@ var _ = Describe("Zcash", func() {
 
 				// Build the transaction by consuming the outputs and spending
 				// them to a set of recipients.
-				inputs := []bitcoincompat.Input{
+				inputs := []utxo.Input{
 					{Output: output},
 				}
-				recipients := []bitcoincompat.Recipient{
+				recipients := []utxo.Recipient{
 					{
-						Address: pack.String(pkhAddr.EncodeAddress()),
-						Value:   pack.NewU64((output.Value.Uint64() - 1000) / 2),
+						To:    address.Address(pkhAddr.EncodeAddress()),
+						Value: pack.NewU256FromU64(pack.NewU64((output.Value.Int().Uint64() - 1000) / 2)),
 					},
 					{
-						Address: pack.String(pkhAddrUncompressed.EncodeAddress()),
-						Value:   pack.NewU64((output.Value.Uint64() - 1000) / 2),
+						To:    address.Address(pkhAddrUncompressed.EncodeAddress()),
+						Value: pack.NewU256FromU64(pack.NewU64((output.Value.Int().Uint64() - 1000) / 2)),
 					},
 				}
-				tx, err := zcash.NewTxBuilder(&chaincfg.RegressionNetParams).BuildTx(inputs, recipients)
+				tx, err := zcash.NewTxBuilder(&zcash.RegressionNetParams, 1000000).BuildTx(inputs, recipients)
 				Expect(err).ToNot(HaveOccurred())
 
 				// Get the digests that need signing from the transaction, and
@@ -90,7 +90,9 @@ var _ = Describe("Zcash", func() {
 
 				// Submit the transaction to the Bitcoin Cash node. Again, this
 				// should be running a la `./multichaindeploy`.
-				txHash, err := client.SubmitTx(context.Background(), tx)
+				txHash, err := tx.Hash()
+				Expect(err).ToNot(HaveOccurred())
+				err = client.SubmitTx(context.Background(), tx)
 				Expect(err).ToNot(HaveOccurred())
 				log.Printf("TXID               %v", txHash)
 
@@ -120,128 +122,6 @@ var _ = Describe("Zcash", func() {
 				Expect(err).ToNot(HaveOccurred())
 				Expect(reflect.DeepEqual(output, output2)).To(BeTrue())
 			})
-		})
-	})
-
-	Context("when sending ZEC to a script", func() {
-		It("should work", func() {
-			// Private key
-			pkEnv := os.Getenv("ZCASH_PK")
-			if pkEnv == "" {
-				panic("ZCASH_PK is undefined")
-			}
-			wif, err := btcutil.DecodeWIF(pkEnv)
-			Expect(err).ToNot(HaveOccurred())
-
-			// Client
-			client := bitcoincompat.NewClient(bitcoincompat.DefaultClientOptions().WithHost("http://127.0.0.1:18232"))
-
-			// Script
-			gpubkey := wif.PrivKey.PubKey().SerializeCompressed()
-			ghash := [32]byte{}
-			gscript, err := bitcoincompat.GatewayScript(gpubkey, ghash)
-			Expect(err).ToNot(HaveOccurred())
-			gaddr, err := zcash.NewAddressScriptHash(gscript, &chaincfg.RegressionNetParams)
-			Expect(err).ToNot(HaveOccurred())
-
-			// PKH
-			pkh, err := zcash.NewAddressPubKeyHash(btcutil.Hash160(wif.PrivKey.PubKey().SerializeCompressed()), &chaincfg.RegressionNetParams)
-			Expect(err).ToNot(HaveOccurred())
-			log.Printf("PKH %v", pkh.EncodeAddress())
-
-			scriptOutput := bitcoincompat.Output{}
-			{
-				// Outputs
-				outputs, err := client.UnspentOutputs(context.Background(), 0, 999999999, pkh)
-				Expect(err).ToNot(HaveOccurred())
-				Expect(len(outputs)).To(BeNumerically(">", 0))
-				output := outputs[0]
-
-				// Input, recipients, and transaction
-				inputs := []bitcoincompat.Input{
-					{Output: output},
-				}
-				recipients := []bitcoincompat.Recipient{
-					{Address: pack.String(gaddr.EncodeAddress()), Value: pack.NewU64(output.Value.Uint64() - 1000)},
-				}
-				tx, err := zcash.NewTxBuilder(&chaincfg.RegressionNetParams).BuildTx(inputs, recipients)
-				Expect(err).ToNot(HaveOccurred())
-
-				// Sign
-				sighashes, err := tx.Sighashes()
-				signatures := make([]pack.Bytes65, len(sighashes))
-				Expect(err).ToNot(HaveOccurred())
-				for i := range sighashes {
-					hash := id.Hash(sighashes[i])
-					privKey := (*id.PrivKey)(wif.PrivKey)
-					signature, err := privKey.Sign(&hash)
-					Expect(err).ToNot(HaveOccurred())
-					signatures[i] = pack.NewBytes65(signature)
-				}
-				Expect(tx.Sign(signatures, pack.NewBytes(wif.SerializePubKey()))).To(Succeed())
-
-				// Submit
-				txHash, err := client.SubmitTx(context.Background(), tx)
-				Expect(err).ToNot(HaveOccurred())
-
-				for {
-					// Confirm
-					confs, err := client.Confirmations(context.Background(), txHash)
-					Expect(err).ToNot(HaveOccurred())
-					log.Printf("%v has %v/3 confirmations", txHash, confs)
-					if confs >= 1 {
-						break
-					}
-					time.Sleep(time.Second)
-				}
-
-				outputs, err = tx.Outputs()
-				Expect(err).ToNot(HaveOccurred())
-				Expect(outputs).To(HaveLen(1))
-				scriptOutput = outputs[0]
-			}
-
-			{
-				// Input, recipients, and transaction
-				inputSigScript, err := bitcoincompat.GatewayScript(gpubkey, ghash)
-				Expect(err).ToNot(HaveOccurred())
-				inputs := []bitcoincompat.Input{
-					{Output: scriptOutput, SigScript: inputSigScript},
-				}
-				recipients := []bitcoincompat.Recipient{
-					{Address: pack.String(pkh.EncodeAddress()), Value: pack.NewU64(scriptOutput.Value.Uint64() - 1000)},
-				}
-				tx, err := zcash.NewTxBuilder(&chaincfg.RegressionNetParams).BuildTx(inputs, recipients)
-				Expect(err).ToNot(HaveOccurred())
-
-				// Sign
-				sighashes, err := tx.Sighashes()
-				signatures := make([]pack.Bytes65, len(sighashes))
-				Expect(err).ToNot(HaveOccurred())
-				for i := range sighashes {
-					hash := id.Hash(sighashes[i])
-					privKey := (*id.PrivKey)(wif.PrivKey)
-					signature, err := privKey.Sign(&hash)
-					Expect(err).ToNot(HaveOccurred())
-					signatures[i] = pack.NewBytes65(signature)
-				}
-				Expect(tx.Sign(signatures, pack.NewBytes(wif.SerializePubKey()))).To(Succeed())
-
-				// Submit
-				txHash, err := client.SubmitTx(context.Background(), tx)
-				Expect(err).ToNot(HaveOccurred())
-
-				for {
-					// Confirm
-					confs, err := client.Confirmations(context.Background(), txHash)
-					Expect(err).ToNot(HaveOccurred())
-					log.Printf("%v has %v/3 confirmations", txHash, confs)
-					if confs >= 3 {
-						break
-					}
-					time.Sleep(time.Second)
-				}
-			}
 		})
 	})
 })
